+++ src/main/kotlin/com/codescene/jetbrains/util/Log.kt
@@ -1,0 +1,32 @@
+package com.codescene.jetbrains.util
+
+import com.codescene.jetbrains.util.Constants.CODESCENE
+import com.intellij.openapi.diagnostic.Logger
+
+/**
+ * Logs are written to `idea.log`, located in the `build/idea-sandbox/system/log/`
+ * directory during plugin development.
+ *
+ * To view logs, configure the log file in your Run/Debug configuration by
+ * specifying the path and alias to display in the console.
+ *
+ * To view debug logs specifically, enable the appropriate debug settings
+ * in the IDE by navigating to Help > Diagnostic Tools > Debug Log Settings
+ * and adding the logger path, e.g., `#com.codescene.jetbrains.util.Log`.
+ * Make sure to filter by all types of logs in the console, not just warnings.
+ *
+ * @see <a href="https://www.jetbrains.com/help/idea/setting-log-options.html#add_log">IntelliJ Logging Documentation</a>
+ */
+
+object Log {
+    private val logger: Logger by lazy { Logger.getInstance(Log::class.java) }
+    private const val PREFIX = "$CODESCENE -"
+
+    fun info(message: String) = logger.info("$PREFIX $message")
+
+    fun warn(message: String) = logger.warn("$PREFIX $message")
+
+    fun debug(message: String) = logger.debug("$PREFIX $message")
+
+    fun error(message: String) = logger.error("$PREFIX $message")
+}